Create createProcess ipc interface
