+++ src/vs/platform/terminal/electron-main/localPtyMainService.ts
@@ -3,8 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { IProcessEnvironment } from 'vs/base/common/platform';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { ICommonLocalPtyService } from 'vs/platform/terminal/common/terminal';
+import { ICommonLocalPtyService, IShellLaunchConfig } from 'vs/platform/terminal/common/terminal';
 
 export const ILocalPtyMainService = createDecorator<ILocalPtyMainService>('localPtyMainService');
 
@@ -13,8 +14,7 @@ export interface ILocalPtyMainService extends ICommonLocalPtyService { }
 export class LocalPtyMainService implements ICommonLocalPtyService {
 	declare readonly _serviceBrand: undefined;
 
-	// TODO: Remove test function
-	async test(): Promise<void> {
+	async createProcess(shellLaunchConfig: IShellLaunchConfig, cwd: string, cols: number, rows: number, env: IProcessEnvironment, windowsEnableConpty: boolean): Promise<void> {
 		console.log('PtyMainService#test');
 	}
 }
